+++ lib/util/async_store.dart
@@ -76,8 +76,10 @@ abstract class _AsyncStore<T> with Store {
     bool refresh = false,
   }) async {
     try {
-      return await run(() => LemmyApiV3(instanceHost).run(query),
-          refresh: refresh);
+      return await run(
+        () => LemmyApiV3(instanceHost).run(query),
+        refresh: refresh,
+      );
     } on LemmyApiException catch (err) {
       final data = refresh ? asyncState.mapOrNull(data: (data) => data) : null;
       if (data != null) {