+++ lib/util/extensions/api.dart
@@ -52,3 +52,14 @@ extension UserPreferredNames on PersonSafe {
 extension CommentLink on Comment {
   String get link => 'https://$instanceHost/post/$postId/comment/$id';
 }
+
+// inspired by https://github.com/LemmyNet/lemmy-ui/blob/66c846ededef8c0afd5aaadca4aaedcbaeab3ee6/src/shared/utils.ts#L533
+extension PersonSafeCakeDay on PersonSafe {
+  bool get isCakeDay {
+    final now = DateTime.now().toUtc();
+
+    return now.day == published.day &&
+        now.month == published.month &&
+        now.year != published.year;
+  }
+}