+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+### Added
+
+- You can now add an instance from the three dots menu on the instance page
+
 ## v0.8.0 - 2022-01-14
 
 ### Added